Created UI Framework v1.00a
This update adds the addition of version control. This is version v1.00a. It includes the addition of a UI without the functionality of the graph portion. The reason for this is that it is being reworked. The previous version graphed a function based on a cameras location, but I learned that, that method would be hard to implement with a UI. So, I am rewriting it so that a graph is calculated on a static position where the bounds are changed. I also have added the basic framework of the functionality of the program that can parse through a users input to decide what to graph.